talkers/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.8.2
bitstring==4.1.2
black==23.10.1
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.7.22
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.0
chinese-converter==1.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-spectacular==0.26.5
executing==1.2.0
filelock==3.12.4
future==0.18.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
inflection==0.5.1
ipykernel==6.25.2
ipython==8.15.0
iso8601==2.1.0
jedi==0.19.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
lxml==4.9.3
lz4==4.3.2
m3u8==3.6.0
matplotlib-inline==0.1.6
mpegdash==0.3.1
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
outcome==1.3.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pathvalidate==3.2.0
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pycryptodomex==3.19.0
Pygments==2.16.1
PyJWT==2.8.0
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
stevedore==5.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.11.1
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.24.0.post1
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
websockets==11.0.3
webvtt-py==0.4.6
wsproto==1.2.0
yt-dlp==2023.10.13
